+++ Makefile
@@ -81,3 +81,30 @@ build-goosey:
 		go build -v -o goosey . && cd $(CUR_DUR)
 	@echo copying goosey binary from $(GOOSEY_PROJ_DIR) to current directory
 	@cp $(GOOSEY_PROJ_DIR)/goosey ./goosey
+
+fetch-tags:
+	@{ \
+	  branch=$$(git rev-parse --abbrev-ref HEAD); \
+	  if [ "$$branch" != "$(MAIN_BRANCH)" ]; then \
+	    echo "Error: You must be on the $(MAIN_BRANCH) branch. Current branch is '$$branch'."; \
+	    exit 1; \
+	  fi; \
+	  git fetch origin $(MAIN_BRANCH); \
+	  UPSTREAM=origin/$(MAIN_BRANCH); \
+	  LOCAL=$$(git rev-parse @); \
+	  REMOTE=$$(git rev-parse "$$UPSTREAM"); \
+	  BASE=$$(git merge-base @ "$$UPSTREAM"); \
+	  if [ "$$LOCAL" != "$$REMOTE" ]; then \
+	    echo "Error: Your local $(MAIN_BRANCH) branch is not up-to-date with remote. Please pull the latest changes."; \
+	    exit 1; \
+	  fi; \
+	  git fetch --tags; \
+	}
+
+release: fetch-tags
+	@{ \
+	  echo "Latest tag: $(LATEST_TAG)"; \
+	  new_tag=$$(go run . utils version-bumper --latest-version "$(LATEST_TAG)" --increment-type=$(VERSION_TYPE)); \
+	  echo "Creating new tag: $$new_tag"; \
+	  git tag -a $$new_tag -m $$new_tag && git push --tags; \
+	}
